FIX unsafe handling of usersDB
usersdb is a module-level object, so if we want that the app behave in a
almost functional way, it must clean the module-level objects. This
still excludes any possible parallel running of two functions (they will
contend the usage of usersDB) but at least non-parallel running of two
flask applications (which is the case of unit tests) will work fine.